Update patch set 1
Patch Set 1:

dropped the others (1 & 2) build was successful without!

ifneq ($(filter 10.6 10.6.%,$(DARWIN_VERSION)),)

The idea is to Force use of SDK 10.5 if we're on Leopard or Snow Leopard (Xcode 3.x)

ifneq ($(filter 10.5 10.6 10.5.% 10.6.%,$(DARWIN_VERSION)),) should do it.
If DARWIN_VERSION < 10.5 an error should have been triggered before we hit this line
if DARWIN_VERSION >= 10.7 we skip  and use whatever is there
is that right ?

Patch-set: 1